requirements.txt
flask==3.1.2
requests==2.32.5
python-dotenv==1.1.1
pillow==11.3.0